Changed players from HashSet to array of Player record
Changed the way that players are stored within the TelepathyGame
class to allow for keeping extra information related to that player
in the future.

When a player attempts to join a game verification is in place to
check there is a spot available and their name is not already taken.