Updated client sendCommand to append list of attributes
The sendCommand method on the client side now can send an array
of strings adding extra information to a CommandPackage for the
server to interpret that is linked to that command.
The working example is the BUTTONPRESS command that can send the
button coordinate as a string.

The TelepathyCommands enum has been updated with new documentation
that describes the format to use for each command.

Tests have been updated to reflect the changes with joining a game.