Telepathy client class implements Tickable. Player connection works correctly.
The Telepathy game can now regularly receive updates from the server
by implementing the Tickable interface. On each tick the client
requests an update from the server. The server is now able to send
the client a series of RenderingCommands in a RenderingPackage with
information the client can use to render the latest state.

Also fixed issues with client/server communications when players leave and
join a game.