Updated unit tests to cover new netcode changes
The unit tests now check TelepathyServer.callGame() to ensure it is correctly
using the TelepathyGame.runCommand() method. Testing also implemented for the QUIT,
SYSTEMQUIT, and invalid commands on the runCommand() method.

runCommand has been modified to handle any command that does evaluate to a
TelepathyCommands enum value by just returning an INVALIDCOMMAND to the sender.

Documentation that was missing has been added to the client and server side methods.