Changed players from array to HashMap
Players are now stored in a HashMap to make it easier to access specific players
and allow for easier implementation of more players at a later stage in the game.
The max number of players allowed in a game is set by maxPlayers field and default
initialised to 2. This could later be changed by players hosting the game.

Tests and other code relating to players has been modified and works with the
HashMap datatype.