Client can send TOGGLEREADY state and state game.
When players first join a game they are waiting for the game to start. It will
not start until the game has the required number of players and they are all
ready to begin playing. The TelepathyGame now has the required functionality to
receive TOGGLEREADY commands and respond to the client with them.