Added updateQueue to Player
The update queue is a list of renderingCommands to be sent to that Player
on the next REQUESTUPDATE tick. When the queue is sent it is cleared and ready to accept new commands. If there are no commands in the queue then a NOUPDATE command is sent.

Added new enum constants to TelepathyCommands to handle changing to GAMEOVER state for clients.