Better handling for sending/receiving commands
Added makeJsonCommand on server side to allow easier creation of RenderingCommands.

Added getAttributes to easily grab the array of attributes associated with a command for parsing RenderingPackages.

Server and Client communication is running well and the INITIALISE state is running as expected allowing players to join and wait until everyone is ready before entering the RUNNING state. The game ends correctly when a player leaves during game play and the server boots all remaining players.