Logic bug fix in Telepathy client class
Created methods to disable and enable all buttons on the board to remove the bug of unwanted popups during game play. Updated col and row references across server and client so that col is always x and row is always y. Removed redundant getTile method on server and updated testGetTile(). Tidied up code on client and added border around player's selected tile...refactoring to be continued.