Update `GameShowPlayer` class
The class representing players within `GameShow.java` on the server has been updated from a record to a private class in order to be able to store mutable states for each player, including scores and ready state.

The 'Start a game' button in the (client) UI has also been updated with a listener which toggles the ready state of the player.