Update GameShow to check all players are ready
GameShow now performs a check when the 'Start a game' button is clicked to determine if all players are ready. Currently, just logs a message but eventually this will be the factor determining when a game begins.

Also updated the client `GameShow.java` to use Log4j.

Also updated the client to log a message when it quits a game.